Reduced # of registers; removed more code; windows emulator and calc.bin are built / untested
git-svn-id: http://svn.code.sf.net/p/wp34s/code@3552 9654ca62-1852-461c-aa49-1be32f3bb02d

+++ branches/wp31s/create_revision.c
@@ -127,7 +127,7 @@ int get_revision_num(char *vcs_cmd, char *vcs)
 	}
 
 	// Convert any LFs to spaces
-	for (i=0; i<strlen(buffer); i++)
+	for (i=0; i<(int)strlen(buffer); i++)
 	{
 	    if (buffer[i] == '\n')
 		buffer[i] = ' ';
